requirements.txt
fastapi==0.121.3
starlette>=0.47.2,<0.51.0
uvicorn[standard]==0.34.0
pillow>=12.1.1,<13.0.0
numpy<2.0
opencv-python-headless==4.10.0.84
onnxruntime==1.19.2
rembg==2.0.57
openai>=2.0.0,<3.0.0
